+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Releases Changelog
 
 ## 1.1.0
+- Results are now delivered as a `zip` file.
 - Added a `PATCH` endpoint for `/registries` so it's easier to update credentials
 - Added the `active` field for registries, so outdated ones can be safely deactivated
 - `cert-manager`'s Certificate now supports `rotationPolicy` via the `certs.rotationPolicy` field. Defaults to `Never`. The other value supported is `Always`.